Move eventListeners to load-list module
